chore(refactor): push analytics
- Refactored push analytics code
