+++ package.json
@@ -29,6 +29,7 @@
     "react-hook-form": "^7.43.9",
     "react-router-dom": "^6.11.1",
     "recoil": "^0.7.7",
-    "usehooks-ts": "^2.9.1"
+    "usehooks-ts": "^2.9.1",
+    "zod": "^3.21.4"
   }
 }
