+++ src/modules/PushAnalytics/index.tsx
@@ -1,10 +1,6 @@
 import React from "react";
 import PushAnalyticsTable from "./components/PushAnalyticsTable";
 
-function anAddPushAnalytics() {
-    console.log("on add push analytics");
-}
-
 export default function PushAnalytics(): React.ReactElement {
 
     return (
